refactor: combine ContainerYoutubeSubscribe with UIYoutubeSubscribe
