+++ components/UIYoutubeSubscribe.vue
@@ -7,6 +7,8 @@
   />
 </template>
 <script>
+// Document: https://developers.google.com/youtube/youtube_subscribe_button
+
 export default {
   name: 'UIYoutubeSubscribe',
   props: {
@@ -15,6 +17,26 @@ export default {
       required: true,
     },
   },
+  mounted() {
+    window.addEventListener('load', this.init)
+  },
+  beforeDestroy() {
+    window.removeEventListener('load', this.init)
+  },
+  methods: {
+    init() {
+      const hasInjectedSDK =
+        window.gapi || document.querySelector('#js-yt-platform')
+      if (!hasInjectedSDK) {
+        const youtubePlatformScript = document.createElement('script')
+        youtubePlatformScript.setAttribute('id', 'js-yt-platform')
+        youtubePlatformScript.setAttribute(
+          'src',
+          'https://apis.google.com/js/platform.js'
+        )
+        document.head.appendChild(youtubePlatformScript)
+      }
+    },
+  },
 }
 </script>
-<style lang="scss" scoped></style>
